update urban soil, water content.
