--- _release/manifest.yml
@@ -1,13 +1,0 @@
----
-name: BeingAliveLanguage
-version: 0.2.1.0
-authors:
-- Dr. Zhao Ma @ BeingAlive
-description: This is the plugin for automatically using the set of language developed by the Chair of Being Alive at ETH Zurich.
-url: ma@arch.ethz.ch
-icon_url: https://i.imgur.com/WABE4LN.png
-keywords:
-  - drawing
-  - climate
-  - soil
-  - language
